% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjustResiduals}
\alias{adjustResiduals}
\title{Make standardized residual adjustments}
\usage{
adjustResiduals(lm_obj)
}
\arguments{
\item{lm_obj}{linear model object}
}
\value{
Adjusted residuals
}
\description{
Make standardized residual adjustments
}
