% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootLMEngine}
\alias{bootLMEngine}
\title{Controls the bootstrapping}
\usage{
bootLMEngine(lm_obj, new_df, seed = 1)
}
\arguments{
\item{lm_obj}{Linear model object}

\item{new_df}{Prediction data}

\item{seed}{[1] random seed}
}
\value{
Cross validation prediction
}
\description{
Controls the bootstrapping
}
