% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateAccuracy}
\alias{calculateAccuracy}
\title{Calculate model accuracy}
\usage{
calculateAccuracy(score_df)
}
\arguments{
\item{score_df}{A scored data.frame}
}
\value{
Dzta.frame of accuracy stats
}
\description{
Calculate standard accuracy metrics
}
