% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateErrors}
\alias{calculateErrors}
\title{Calculate the prediction error on the test set}
\usage{
calculateErrors(train_df)
}
\arguments{
\item{test_df}{Data.frame of data to score}

\item{point_df}{Data.frame with point prediction}
}
\value{
Test observation with prediction error
}
\description{
Calculate the prediction error on the test set
}
