% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateFSD}
\alias{calculateFSD}
\title{Calculate the FSD of test set}
\usage{
calculateFSD(train_df, pred_intervals, pi_methods)
}
\arguments{
\item{pred_intervals}{Intervals to calculate bounds at}

\item{pi_methods}{Methods to use (if 'error' then do, else return NULL)}

\item{test_df}{Data.frame of data to score}
}
\value{
List with FSD and FSD bounds
}
\description{
Includes the various normal distribution expansion at the various prediction intervals
}
