% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calibEfficiency}
\alias{calibEfficiency}
\title{Calculate calibration for a scored object}
\usage{
calibEfficiency(interval_df, intervals)
}
\arguments{
\item{scored_obj}{An object from scoreModel}
}
\value{
List of model and error calib and intervals
}
\description{
Calculate calibration for a scored object
}
