% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calibEngine}
\alias{calibEngine}
\title{Calculate calibration for a scored object}
\usage{
calibEngine(within_df)
}
\arguments{
\item{within_df}{A checkPI completed scored dateaset}
}
\value{
List of model and error calib and intervals
}
\description{
Calculate calibration for a scored object
}
