% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calibrateModel}
\alias{calibrateModel}
\title{Calculate calibration for a scored object}
\usage{
calibrateModel(scored_obj)
}
\arguments{
\item{scored_obj}{An object from scoreModel}
}
\value{
List of model and error calib and intervals
}
\description{
Calculate calibration for a scored object
}
