% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkPredIntervals}
\alias{checkPredIntervals}
\title{Determine whether or not predictions are within the PI}
\usage{
checkPredIntervals(interval_df)
}
\arguments{
\item{interval_df}{A prediction interval data.frame ()}
}
\value{
Dzta.frame with 0/1 withinness for each PI
}
\description{
Runs a validation run given an information horizon, produces results
}
