% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crossVal}
\alias{crossVal}
\title{Cross validation for linear model}
\usage{
crossVal(train_df, mod_spec, k = 5)
}
\arguments{
\item{train_df}{Training data}

\item{mod_spec}{Model specification}

\item{k}{[5] Number of folds}
}
\value{
Cross validation prediction
}
\description{
Cross validation for linear model
}
