% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{depthInterval}
\alias{depthInterval}
\title{Calculate the depth-related intervals}
\usage{
depthInterval(x_df, fsd, intervals)
}
\arguments{
\item{x_df}{Method to use for interval construction}

\item{fsd}{Depth specific FSD}

\item{intervals}{Prediction intervals}
}
\value{
Prediction df from depth
}
\description{
Calculate the depth-related intervals
}
