% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractIntervals}
\alias{extractIntervals}
\title{Extract raw PI values}
\usage{
extractIntervals(raw_obj, interval)
}
\arguments{
\item{raw_obj}{Raw data object}

\item{interval}{Interval at which to extract the raw PIs}
}
\value{
Raw intervals
}
\description{
Extract raw PI values
}
