% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractResults}
\alias{extractResults}
\title{Extracts and summarize all data from raw experiment output}
\usage{
extractResults(file_name, intervals = NULL)
}
\arguments{
\item{file_name}{Raw data}

\item{intervals}{[NULL] Vector of interval levels at which to extract the raw results}
}
\value{
Summarized extracted data
}
\description{
Extracts and summarize all data from raw experiment output
}
