% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotEffic}
\alias{plotEffic}
\title{Validation of a training model/data}
\usage{
plotEffic(cal_obj, title_prefix = "Calibration", xaxis_gap = 0.1,
  yaxis_gap = 0.1)
}
\arguments{
\item{cal_obj}{object from calibrateModel()}

\item{title_prefix}{['Calibration'] Title prefix}

\item{xaxis_gap}{[.1] Gap for x axis}

\item{yaxis_gap}{[.1] Gap for y axis}
}
\value{
plot of calibration comparisons
}
\description{
Runs a validation run given an information horizon, produces results
}
