% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotQuantiles}
\alias{plotQuantiles}
\title{Plot pi quantiles vs errors cumulatives}
\usage{
plotQuantiles(score_obj, plot_title = "Insert Title Here",
  pi_width = 50)
}
\arguments{
\item{score_obj}{object from calibrateModel()}

\item{plot_title}{['Insert Title Here'] Title}

\item{pi_width}{[50] PI width to examine in the Title prefix}
}
\value{
plot of error vs pi quantiles
}
\description{
Plot pi quantiles vs errors cumulatives
}
