% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pointPrediction.rf}
\alias{pointPrediction.rf}
\title{Point Prediction from a random forest model}
\usage{
\method{pointPrediction}{rf}(mod_obj, test_df, pred_intervals, ...)
}
\arguments{
\item{mod_obj}{A model object from trainModel()}

\item{test_df}{Data.frame of data to score}

\item{pred_intervals}{[.05 to .95 by .05] Desired prediction intervals}

\item{...}{Additional arguments}
}
\description{
Point Prediction from a random forest model
}
