% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rangePrediction}
\alias{rangePrediction}
\title{Interval scoring for model based PIs}
\usage{
rangePrediction(mod_obj, test_df, pred_intervals, ...)
}
\arguments{
\item{mod_obj}{Trained model objects}

\item{test_df}{Data.frame of data to score}

\item{pred_intervals}{[.05 to .95 by .05] Desired prediction intervals}

\item{...}{Additional arguments}
}
\value{
Data.frame of prediction intervals
}
\description{
Generic method for creating prediction intervals from a model based method
}
