% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rangePrediction.rf}
\alias{rangePrediction.rf}
\title{Score intervals with random forest}
\usage{
\method{rangePrediction}{rf}(mod_obj, test_df, pred_intervals, ...)
}
\arguments{
\item{mod_obj}{Trained model objects}

\item{test_df}{Data.frame of data to score}

\item{pred_intervals}{[.05 to .95 by .05] Desired prediction intervals}

\item{...}{Additional arguments}
}
\description{
Score intervals with random forest
}
