% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rollingPreds}
\alias{rollingPreds}
\title{Wrapper to roll through a number of dates and do foreward casts}
\usage{
rollingPreds(data_df, roll_dates, model_type, model_spec,
  keep_results = FALSE, ...)
}
\arguments{
\item{data_df}{Raw data}

\item{roll_dates}{Datest to evaluate model on}
}
\value{
Test data and model calibrations
}
\description{
Wrapper to roll through a number of dates and do foreward casts
}
