% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scoreInterval.model}
\alias{scoreInterval.model}
\title{Score intervals with model method}
\usage{
\method{scoreInterval}{model}(pi_method, pred_intervals, test_df, mod_obj,
  fsd_obj = NULL, ...)
}
\arguments{
\item{pi_method}{Method to use for interval construction}

\item{pred_intervals}{[.05 to .95 by .05] Desired prediction intervals}

\item{test_df}{Data.frame of data to score}

\item{mod_obj}{Trained model objects}

\item{fsd_obj}{[NULL] fsb object from calculateFSD()}

\item{...}{Additional arguments}
}
\description{
Score intervals with model method
}
