% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scoreModel}
\alias{scoreModel}
\title{Score models}
\usage{
scoreModel(mod_obj, test_df, pi_methods = c("model", "error"),
  pred_intervals = seq(0.05, 0.95, by = 0.05), ...)
}
\arguments{
\item{mod_obj}{A model object from trainModel()}

\item{test_df}{Data.frame of data to score}

\item{pi_methods}{[c('model', 'error', 'instance')] Different methods for producing pred intervals}

\item{pred_intervals}{[.05 to .95 by .05] Desired prediction intervals}

\item{...}{Additional arguments}
}
\value{
Scored model object
}
\description{
Wrapper for scoring any model class with any set of pi methods and pred intervals
}
