% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{submarket_df}
\alias{submarket_df}
\title{Custom mapping of assessor areas to 'submarkets'.}
\format{A data.frame (tbl)}
\source{
Manually put together with code in /scripts/create_project_data.R
}
\usage{
submarket_df
}
\description{
Custom mapping of assessor areas to 'submarkets'.
}
\keyword{datasets}
