% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timeAdjustData}
\alias{timeAdjustData}
\title{Time adjust data to current time}
\usage{
timeAdjustData(trans_df)
}
\arguments{
\item{trans_df}{Data.frame of transactions (from kingCoSales)}
}
\value{
List of training and test data
}
\description{
Time adjust data to current time
}
