% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trainModel}
\alias{trainModel}
\title{Train models}
\usage{
trainModel(train_df, mod_type, mod_spec, ...)
}
\arguments{
\item{train_df}{Training data}

\item{mod_type}{Type of model to train}

\item{mod_spec}{Model specification}

\item{...}{Additional arguments}
}
\value{
Trained model object
}
\description{
Generic method for training models
}
