% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trainModel.rf}
\alias{trainModel.rf}
\title{Train a random forest model}
\usage{
\method{trainModel}{rf}(train_df, mod_type, mod_spec, ...)
}
\arguments{
\item{train_df}{Training data}

\item{mod_type}{Type of model to train}

\item{mod_spec}{Model specification}

\item{...}{Additional arguments}
}
\value{
Trained linear model plut optional boot strapped approach
}
\description{
Train a random forest model
}
