% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimPriceQuantiles}
\alias{trimPriceQuantiles}
\title{Create train/test split}
\usage{
trimPriceQuantiles(x_df, lo_qtl, hi_qtl)
}
\arguments{
\item{x_df}{Data.frame to trim from}

\item{lo_qtl}{Lower quantile cutoff}

\item{hi_qtl}{Higher quantile cutoff}
}
\value{
Filtered sales data
}
\description{
Split data based on a horizon ate
}
